+++ src/arguments.rs
@@ -5,7 +5,6 @@ use dsh_api::platform::DshPlatform;
 pub(crate) const APP_ID_ARGUMENT: &str = "app-id-argument";
 pub(crate) const BUCKET_ID_ARGUMENT: &str = "bucket-id-argument";
 pub(crate) const CERTIFICATE_ID_ARGUMENT: &str = "certificate-id-argument";
-#[cfg(feature = "appcatalog")]
 pub(crate) const MANIFEST_ID_ARGUMENT: &str = "manifest-id-argument";
 pub(crate) const PLATFORM_NAME_ARGUMENT: &str = "platform-name-argument";
 pub(crate) const PROXY_ID_ARGUMENT: &str = "proxy-argument";
@@ -45,7 +44,6 @@ pub(crate) fn certificate_id_argument() -> Arg {
     .long_help("Identifies a certificate on the DSH.")
 }
 
-#[cfg(feature = "appcatalog")]
 pub(crate) fn manifest_id_argument() -> Arg {
   Arg::new(MANIFEST_ID_ARGUMENT)
     .action(ArgAction::Set)